requirements.txt
Flask>=2.0
face_recognition>=1.3.0
opencv-python>=4.5
numpy>=1.19
Pillow>=8.0

waitress
